Split the replay collectors into their own files
stored_block_replay_expander.rb was the longest file under project/ at 367
lines, and two of its three classes were visitors doing the recognition rather
than the rewrite. Follow the convention already in place for
inference/class_member_collector/: a directory named after the parent, one
class per file, required from analyzer.rb.

`ReaderCollector` moves out of `Collector` to sit beside it — it touches none
of `Collector`'s state, so nesting bought nothing. The parent file keeps the
rewrite (`expand`/`apply_replays`/indent handling) at 94 lines.

Pure move: no behavior change, snapshots and unit specs unchanged.
